boilerplates/python/requirements.txt
flask
flask-cors
logzero
python-dotenv